% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_P.R
\name{example_P}
\alias{example_P}
\title{Examples of Pokemon With A Given Type}
\usage{
example_P(input)
}
\arguments{
\item{input}{write the "type" for which you want examples}
}
\value{
3 Pokemon names that have your inputted type
}
\description{
The type function takes your own type as input, e.g. type("fire")
and then returns a list of 3 randomly chosen Pokemon
that have that type
}
