% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\title{Recommendation based on opponent's Pokemon's name}
\usage{
name(input)
}
\arguments{
\item{input}{write the "name" of your opponents Pokemon}
}
\value{
types that are strong against your opponents Pokemons type
}
\description{
The name function takes the opponent's Pokemon's name as input,
e.g. type("Pikachu") and then returns a list of recommended counter types
either based on strong defence or strong attack matches
}
