% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{share}
\alias{share}
\title{Share of Pokemon With A Given Type}
\usage{
share(input)
}
\arguments{
\item{input}{write the "type" for which you want the share of Pokemon
that have this type}
}
\value{
the percentage of Pokemon who have the inputted type
}
\description{
The share function takes your own type as input, e.g. type("fire")
and then returns a percentage of how common this type is, e.g.
how many \% of Pokemon have this type
either based on strong defence or strong attack matches
}
