% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type}
\alias{type}
\title{Recommendation based on opponent's type}
\usage{
type(input)
}
\arguments{
\item{input}{write the "type" of your opponents Pokemon}
}
\value{
types that are strong against your inputted type
}
\description{
The type function takes the opponent's type as input, e.g. type("fire")
and then returns a list of recommended counter types
either based on strong defence or strong attack matches
}
